pyproject.toml
[tool.poetry]
name = "squad-schedule"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
djangorestframework-camel-case = "^1.3.0"
django-cors-headers = "^3.12.0"
dj-database-url = "^0.5.0"
gunicorn = {extras = ["gevent"], version = "^20.1.0"}
humanize = "^4.1.0"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"
whitenoise = "^6.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']